+++ src/graphql/schema/create_entities/create_auth_service/start_authentication/mod.rs
@@ -81,7 +81,7 @@ impl ServiceSchemaBuilder {
                         registration_state: user.registration_state,
                         passkey: user.passkey,
                         authentication_state: Some(auth_state),
-                        id: user.id,
+                        uuid: user.uuid,
                     };
 
                     let updated =