+++ src/utils/clean_string/mod.rs
@@ -1,0 +1,3 @@
+pub fn clean_string(v: &String) -> String {
+    v.replace("\"", "")
+}